+++ ts-sdk/fusionamm-sdk/src/index.ts
@@ -19,3 +19,4 @@ export * from "./pool";
 export * from "./position";
 export * from "./swap";
 export * from "./collectProtocolFees";
+export * from "./resetPoolPrice";