+++ ts-sdk/client/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @crypticdot/fusionamm-client
 
+## 1.0.57
+
 ## 1.0.56
 
 ## 1.0.55