+++ ts-sdk/fusionamm-sdk/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @crypticdot/fusionamm
 
+## 1.0.57
+
+### Patch Changes
+
+- 01c5a9a: Added reset_pool_price instruction
+  - @crypticdot/fusionamm-client@1.0.57
+  - @crypticdot/fusionamm-core@1.0.57
+
 ## 1.0.56
 
 ### Patch Changes